// artifacts/rcx-supervisor/client/src/proto/eag/components/Stats/constant.ts
export const testIds = {
    STATS_SETTINGS_ICON: 'stats-settings-icon',
};
export enum DashboardType {
    DAILY_METRICS = 'DailyMetrics',
    INBOUND_METRICS = 'InboundMetrics',
    OUTBOUND_METRICS = 'OutboundMetrics',
    BLENDED_METRICS = 'BlendedMetrics',
    CHAT_METRICS = 'ChatMetrics',
}
export enum StatsTableColumnType {
    STRING = 'string',
    NUMBER = 'number',
    DURATION = 'duration',
    PERCENT = 'percent',
}

export const TOTAL_SAFE_DIALS = 'totalSafeDials';
export const TOTAL_PREVIEW_DIALS = 'totalPreviewDials';